data-space: segment the viewer screen
