+++ packages/legend-application-taxonomy/style/_taxonomy-explorer.scss
@@ -135,7 +135,7 @@
   }
 }
 
-.taxonomy-explorer__main-panel {
+.taxonomy-explorer__panel {
   &__header {
     background: var(--color-dark-grey-100);
     z-index: 1;
